healthbank_dev/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.3
django-picklefield==1.0.0
django-q==1.0.0
djangorestframework==3.7.7
lxml==4.2.3
Pillow==5.1.0
pipreqs==0.4.9
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyfcm==1.4.5
pyquery==1.4.0
python-dateutil==2.7.3
pytz==2018.3
requests==2.18.4
requests-html==0.9.0
requests-toolbelt==0.8.0
SQLAlchemy==1.2.10
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
websockets==5.0.1